backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy>=2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0
python-multipart==0.0.9
anthropic==0.28.0
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
psycopg2-binary==2.9.9
